# Data API: Constants and Models are now loaded from the Resource Definition rather than Dynamically