--- a/data/Pandora.Data/Transformers/VersionTests.cs
+++ b/data/Pandora.Data/Transformers/VersionTests.cs
@@ -72,7 +72,7 @@ public static class VersionTests
             typeof(SomeObject),
         };
     }
-    
+
     private class SomeObject
     {
     }