… data/Pandora.Data/Transformers/Constant.cs
… data/Pandora.Data/Transformers/ServiceTests.cs
… data/Pandora.Definitions/Interfaces/ResourceDefinition.cs